+++ examples/llamadeploy/workflow/ui/next.config.ts
@@ -1,0 +1,8 @@
+import type { NextConfig } from 'next'
+
+const nextConfig: NextConfig = {
+  // The base path to create proxy to nextjs frontend in LlamaDeploy
+  basePath: '/deployments/workflow/ui',
+}
+
+export default nextConfig