+++ packages/chat-ui/src/chat/canvas/artifacts/document.tsx
@@ -14,11 +14,31 @@ interface DocumentArtifactViewerProps {
   children?: React.ReactNode
 }
 
+// Convert citation with node_id to citation number in markdown
+function processDocument(content: string, nodeIds: string[]) {
+  if (nodeIds.length === 0) return content
+
+  const citationRegex =
+    /\[citation:([a-fA-F0-9\\-]+)\](?:\(javascript:void\(0\)\))?/g
+
+  let processedContent = content.replace(citationRegex, (match, citationId) => {
+    const nodeIndex = nodeIds.findIndex(nodeId => nodeId === citationId)
+    if (nodeIndex !== -1) return ` \`${nodeIndex + 1}\` `
+    return match // return original citation if not found
+  })
+
+  // MdxEditor does not support <br> tags
+  processedContent = processedContent.replace(/<br\s*\/?>/gi, ' ')
+
+  return processedContent
+}
+
 export function DocumentArtifactViewer({
   className,
   children,
 }: DocumentArtifactViewerProps) {
   const { displayedArtifact, updateArtifact } = useChatCanvas()
+
   const [updatedContent, setUpdatedContent] = useState<string | undefined>()
 
   if (displayedArtifact?.type !== 'document') return null
@@ -27,6 +47,11 @@ export function DocumentArtifactViewer({
   const {
     data: { content, title, type },
   } = documentArtifact
+
+  const transformedContent = processDocument(
+    content,
+    documentArtifact.data.sources?.map(source => source.id) ?? []
+  )
 
   const handleDocumentChange = (markdown: string) => {
     setUpdatedContent(markdown)
@@ -50,9 +75,9 @@ export function DocumentArtifactViewer({
         </h3>
         <ChatCanvasActions />
       </div>
-      <div className="relative mx-20 flex min-h-0 flex-1 flex-col items-stretch gap-4 py-4">
+      <div className="relative mx-20 flex min-h-0 flex-1 flex-col items-stretch gap-4 py-2">
         {updatedContent && (
-          <div className="bg-background absolute right-0 top-2 flex gap-2 py-2 pr-2 text-sm">
+          <div className="absolute right-[30px] top-[14px] z-20 flex gap-2 text-sm">
             <Button
               size="sm"
               className="h-7 bg-blue-500 hover:bg-blue-600"
@@ -73,8 +98,9 @@ export function DocumentArtifactViewer({
         {children ?? (
           <DocumentEditor
             key={documentArtifact.created_at}
-            content={content}
+            content={transformedContent}
             onChange={handleDocumentChange}
+            className="overflow-y-auto"
           />
         )}
       </div>