+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,86 @@
 # nextjs-chat-ui-example
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [db881e1]
+  - @llamaindex/chat-ui@0.5.13
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies [f28cb7e]
+  - @llamaindex/chat-ui@0.5.12
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [05dbe25]
+- Updated dependencies [84a67d3]
+  - @llamaindex/chat-ui@0.5.11
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [b15d152]
+  - @llamaindex/chat-ui@0.5.10
+
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [fe679c4]
+  - @llamaindex/chat-ui@0.5.9
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [d544d57]
+  - @llamaindex/chat-ui@0.5.8
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [23093ed]
+- Updated dependencies [0c929c4]
+  - @llamaindex/chat-ui@0.5.7
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [3c477d5]
+- Updated dependencies [c782810]
+- Updated dependencies [dad784b]
+  - @llamaindex/chat-ui@0.5.6
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [4973f3a]
+  - @llamaindex/chat-ui@0.5.5
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [77620ee]
+  - @llamaindex/chat-ui@0.5.4
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [a34688a]
+  - @llamaindex/chat-ui@0.5.3
+
 ## 0.1.5
 
 ### Patch Changes